📚 Build context and .NET tasks
