+++ docs/articles/recipe/common.md
@@ -1,3 +1,17 @@
 # Common tasks
 
 TODO
+
+## Tasks
+
+### Restore tools
+
+TODO
+
+### Clean artifacts
+
+TODO
+
+### Set version from Git
+
+TODO
