--- docs/articles/workflows/images/translation_workflow.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2021-06-11T13:02:49.515Z" agent="5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.6.13 Chrome/89.0.4389.128 Electron/12.0.7 Safari/537.36" etag="5_0nvXe1aQyriJnV-RWt" version="14.6.13" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7VzbdqM2FP0aP06WDcbxPOY202kzU7dJO81TlwwyqAFEhfBlvr46uoBBTGJnYkO98hLDsRDS2drnJjkD9ypZf2Qoiz7TAMcDZxisB+71wHFGo6EjPkCy0ZLhcKgkISOBllWCO/INm4ZaWpAA57WGnNKYk6wu9GmaYp/XZIgxuqo3W9C4/tYMhdgS3PkotqVfScAjJZ0655X8J0zCyLx5NHmvvkmQaaxnkkcooKstkXszcK8YpVxdJesrHIP2jF6+ftp8jW8fJx9//i3/F/1x+cv9lz/fqc4+7PNIOQWGU/7irm8WfPyQTh/+uZyd3324/YWM8uU711V9L1FcaIUNnEks3nI5FxchXBhBYQQzysVACJK953kB8F6Uj7Hmc7ZEDLJokVlvDMjSGkTclMSkKckzlDZlZaMrmi5iIhaaGHsKi5GhNI8RJxTuGM7oAND9IP7OxVd+tDVA601C1vYyIbZHte/IPyUZzXMyFwtZDAX7qMjhKsDLHKiRBuLvVzwXQwcxDggHeUAWCyzXieAKiXF+Ji6uPsE3FEODlHLdXPay4RFJw7OXTAhQfG00EsxCmM7s13tpKORlzxCAAVn6F9yUywcJQ6kHHok28lHVViieLMovzbKjDEDhEZZd+ygHkLrB4iLLpNmDMYZYQeLTJBPcUCpQrAADbK5Vgxhz3C1QMNgVkeZugeJ4jvxHo+icFsyXuhW+DYR4zYETM4aXBK8A2YLIKcnmYACY7A6xVDAjL+eKhTNieWdMuWckDDEDP2CNfEWEqxLIpPHGrE9frL6wtpaOikhFjKzII6PcBEmS53At7Q9Xk5J2Sc6omlsKPjneSKgqFupu0kI6H933gtGkRsYFBT2RJBFmTkm0prqB7ie6qmhvUESgCI60cakDaeQfOM45DjrlllxN0pOIz5gscek4I7jMUYJbWAY0yomwbhLAO9XI9CLMaFCiVCGu7tXcO6PZ7zhjNCh8sgVMXiFzHBy61/mLlfuktu+ppu9GjjpJVMxCW6xa3sZrqQMk3VQ12WeU3xp9ts3uNY1eHtFCEjr/t0DSRlWz1YbN2MEErUlSJNJ845oeMmknAvDHV8oxz5USfBFmyHBBtCA0OCRX2myYcKAMyZBNRWkapuZylGBtqQSzJUzoFTG4KQeih2DMqHnXEMngSqSbVciM0yUMLpfNeJEpVrGXrP7vCWtpi8wc+cakoxB+CHnEk1gIRuISxSRMxXWMF/CNGDgnIn290OKEBBBFupcM5+Sb9Avu9VDcZ5SkXCbH3uXAu4aeCi44IRPwUfnuZ7NGnV3Ce/F6K2fWWeRHTBPMgWZD/e1Y58O6InA+9NT9qkqvz02RINpKrZ2pFiKd0odl11XaKi505rpHFjt6b2Wxt1ToUOhL53Uryh4XMRQTmoDkK5LEKMV1UPxIrKRbtKEFDF7YR//R3Ckg8MzoddQQfUZrjZB4jHFdEXGGx0PE3G4DMmkBZDI+FCCmZLIFyDVe2iZZomQ8wOz3HcHZ1qvbrtenl8nzitWadGxNOm1Le+IcbGl7llJwEOI7fStylYiGNEXxTSW9ZLRIAxzoZVi1uaU00zr8B3O+0ToEu1HXMF4T/hc8fubpu4etb67NApc3G3OTivluPQS3D9vfVY/JO/PcjuCVVTjlZp7QmC4yIgZJ7BMdqnagzSeXAsNQHVrWy4mvj/OkE5x7gpmzI2ZurzCza6dgwlS1TQhJUMs9dfwBKZ0FdQUkoLKKCMd3GZIqWzGUNQKGeU7jguML5t9V3h5Vd2NpIxl9LCvezv6Yvdj7tBnNkdNiNA/nfSxcrkTErGpUto85Jd2PWlz/cXU/snQvHD+OKcTZqsSzEFAUUC09YRwmXcPg2DiUBc9mEPZZ19xlchcYtPoSia0bYdf3IzO3NcY9WGQ2Hp2ixzaTes5ju+e7hlk989nnFjEsGE0KfovmOJ5BmQmSSPd6TjmnSUuOzmnDPOURyqCzZB3CtvoZpJ5+JLhyRtN3sOX8N8Nymw503KwA7AnV7hwyO9uaQtMWCk1tBhnZq0PhdUOgHqY2u5LOcOlZ0k16xTnHDpRvknm5r6GqdGi7tmtqiT0OEfZmn7NDqOZOjhgjGGJvwXLZKKDuXLz0hbbEA4ctX74SEG7dDHo2DtOWQOJgcYTnvpnB/czgaNcSz/h9r+zgqCUoL+YxkRtEyvSpGoHaV5FmsdoOEtEDWSCf2yXsUzKL7rBrs+jZAeEbH5/k42RHOpr13xM6Tiw2fsQpZrpQtL21qbjZ2OA8YRK2lvCOS8JuyuH/YxIaX/csCz3vtVkoH71gTJ4cNA10fFf1PANBteSm09qSG48bp5cbzUfnP9belN++196ZPNleXKgJVku61NQPlJDsfesv5pRodXJiSeCs2MXs00nbHLelHHFcm+OMLTQ+144jvxs4E5SAWvXZDpm+nhYmfXMEru2l779/aFRvtZ0yJF73dQPPgqTcVDB7Cl/MIbfWH3coEqXzHD7UybiBOWatz/QqxpVt1V/NuO3DxKcM9bhtG+m4UNsm8V7/cOHkIuDeuSObZmVwUPv1yEnD0H05YGxp+C0TeXqXQmvs+a3BV89EfmxDd/qG9J6bwN6OSLv92pAan78hvSfSu3La7Rmn3+pI+1rvnTk97hfS9pHzsoTSWAGnsY3pOPWKVcs+ZtuBqMPtY9rx6jVdpTFFwaBRTwdU4BRslbmfcgzb/U4/PN+A5k7/xLR5GHCm0u9Q/TwXhfDUX2cPFkJdnwf0Jo1KVee/1PDs4xSN3d0eL3JnX/3vcvD4uHna+zdPX8PyWU/v7XqE06zsnnh6zz7C+f8/TvajBDxgqVDcVv/kSG19Vf8ryr35Dw==</diagram></mxfile>
