+++ docs/articles/workflows/translation.md
@@ -1,3 +1,15 @@
 # Translation workflow
 
-![workflow](images/translation_workflow.png)
+A _translation workflow_ would allow to translate applications during the
+development lifecycle in an _agile_ way.
+
+As the developers code the software and _hard-code_ strings in a source language
+(usually English), translators could translate into several languages
+translation files and contribute back to the repository, so every build includes
+the translations as well.
+
+![workflow](images/translation_workflow.drawio.png)
+
+> [!NOTE]  
+> This is an initial design of how a translation workflow could work. It has not
+> been tested yet.
